requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
groq==0.9.0
pinecone==5.4.2
beautifulsoup4==4.12.3
requests==2.32.3
python-dotenv==1.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.32
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pypdf==4.3.1
pydantic-settings==2.4.0
pydantic[email]==2.8.2
httpx==0.27.0
aiofiles==24.1.0